Removed compiler warnings.
git-svn-id: https://www.taskjuggler.org/svn/taskjuggler/trunk@321 e1914e07-63f8-0310-9059-d6d858d7cdca

+++ taskjuggler/Task.cpp
@@ -1864,7 +1864,7 @@ void Task::loadFromXML( QDomElement& parent, Project *project )
       else if( elemTagName == "Priority" )
         setPriority( elem.text().toInt() );
       else if( elemTagName == "complete" )
-	 setComplete( elem.text().toDouble() );
+	 setComplete( elem.text().toInt() );
 
       /* time-stuff: */
       else if( elemTagName == "minStart" )
